Upgrade traefik image to v1.7.5
Upgrade the traefik image to v.1.7.5. This required removing the label
traefik.frontend.passTLSCert=true from the gw-astronomy aa service
because the label has been deprecated. This is not an issue
because we were not using the label since the SPs are expected to
sign SAML attribute authority requests instead of using certificate
authentication.