Add support for managing LetsEncrypt certificates
Add support for managing LetsEncrypt certificates. The webgateway_getssl
service uses supervisor to run 3 processes: Apache HTTP Server, cron,
and busybox syslogd to aid in logging for cron. It also includes getssl
for requesting and renewing certificates. The renewal is run once a day
using cron. The updated Traefik configuration allows the certificates to
be read from the mounted volume and it watches for changes so the
certificates can be updated on the fly.