Update cilogon-web to v28 and oa4mp-demo to v4
The redirects in the cilogon-web image were not working as
expected. The cilogon-web and oa4mp-server services are in separate
containers. In order to accomodate the Virtual Organization
rewrite rules for .well-known, a 302 redirect was needed. This causes
problems with OIDC clients who do not expect a 302. So the redirect
rules have been moved to Traefik replacepathregex rules so that the
oa4mp-server instance handles the requests.

Also, oa4mp-demo had been updated with a custom /etc/resolv.conf file.
Since I figured out the issue, this is no longer necessary.