Implement weighted round robin for cilogon-web.
By using weighted round robin for the cilogon-web service, we avoid the
situation where ECP clients attempt to connect to different instances of the
service. Since some ECP clients do not respsect the sticky session cookie,
the weight for the first service is very high (2^31-1), thus it is very
likely that the ECP client will connect to the same replica every time.
NOTE: This is half the solution. The other half is an updated
tls_certificates.toml file in the webgateway-configs repo.