Update to OA4MP v5.3.1 mk2
Update oa4mp-demo to v18.
Update oa4mp-server to v37.

Jeff found an issue in the last v5.3.1 release requiring an update.
"I did find an issue which I fixed in the release. This is an unchecked
cast for a pretty printing a log message as per CIL-1722 (which should clean
up the logs quite a bit). This has to be fixed to catch weird errors, so I
did it."